WaveRNN

WaveRNN model based on the implementation from fatchord. The original implementation was introduced in "Efficient Neural Audio Synthesis". #' Pass the input through the WaveRNN model.

model_wavernn(
  upsample_scales,
  n_classes,
  hop_length,
  n_res_block = 10,
  n_rnn = 512,
  n_fc = 512,
  kernel_size = 5,
  n_freq = 128,
  n_hidden = 128,
  n_output = 128
)

Arguments

upsample_scales
the list of upsample scales.

n_classes
the number of output classes.

hop_length
the number of samples between the starts of consecutive frames.

n_res_block
the number of ResBlock in stack. (Default: 10)

n_rnn
the dimension of RNN layer. (Default: 512)

n_fc
the dimension of fully connected layer. (Default: 512)

kernel_size
the number of kernel size in the first Conv1d layer. (Default: 5)

n_freq
the number of bins in a spectrogram. (Default: 128)

n_hidden
the number of hidden dimensions of resblock. (Default: 128)

n_output
the number of output dimensions of melresnet. (Default: 128)

Value

Tensor shape: (n_batch, 1, (n_time - kernel_size + 1) * hop_length, n_classes)

Details

forward param:

waveform the input waveform to the WaveRNN layer (n_batch, 1, (n_time - kernel_size + 1) * hop_length)

specgram the input spectrogram to the WaveRNN layer (n_batch, 1, n_freq, n_time)

The input channels of waveform and spectrogram have to be 1. The product of upsample_scales must equal hop_length.

Examples

wavernn = model_wavernn(upsample_scales=c(2,2,5), n_classes=512, hop_length=20)
waveform = torch::torch_rand(3,1,20)
# waveform shape:  (n_batch, n_channel, (n_time - kernel_size + 1) * hop_length)
specgram = transform_mel_spectrogram(16000)(waveform)  # shape: (n_batch, n_channel, n_freq, n_time)

#> Error in (function (self, padding) {    .Call("_torch_cpp_torch_namespace_reflection_pad1d_self_Tensor_padding_IntArrayRef",         PACKAGE = "torchpkg", self, padding)})(self = <pointer: 0x55d1cface170>, padding = c(200L, 200L)): Argument #4: Padding size should be less than the corresponding input dimension, but got: padding (200, 200) at dimension 2 of input [1, 3, 20]
#> Exception raised from reflection_pad1d_out_template at ../aten/src/ATen/native/ReflectionPad.cpp:85 (most recent call first):
#> frame #0: c10::Error::Error(c10::SourceLocation, std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> >) + 0x69 (0x7f904de51b89 in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/deps/./libc10.so)
#> frame #1: <unknown function> + 0xf684d9 (0x7f90385af4d9 in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/deps/./libtorch_cpu.so)
#> frame #2: at::native::reflection_pad1d_cpu(at::Tensor const&, c10::ArrayRef<long>) + 0x12e (0x7f90385b06de in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/deps/./libtorch_cpu.so)
#> frame #3: <unknown function> + 0x13368ed (0x7f903897d8ed in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/deps/./libtorch_cpu.so)
#> frame #4: <unknown function> + 0xaf0fe7 (0x7f9038137fe7 in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/deps/./libtorch_cpu.so)
#> frame #5: <unknown function> + 0x14db322 (0x7f9038b22322 in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/deps/./libtorch_cpu.so)
#> frame #6: at::reflection_pad1d(at::Tensor const&, c10::ArrayRef<long>) + 0xbe (0x7f9038a18d2e in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/deps/./libtorch_cpu.so)
#> frame #7: <unknown function> + 0x298af4b (0x7f9039fd1f4b in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/deps/./libtorch_cpu.so)
#> frame #8: <unknown function> + 0xaf0fe7 (0x7f9038137fe7 in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/deps/./libtorch_cpu.so)
#> frame #9: <unknown function> + 0x14db322 (0x7f9038b22322 in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/deps/./libtorch_cpu.so)
#> frame #10: at::reflection_pad1d(at::Tensor const&, c10::ArrayRef<long>) + 0xbe (0x7f9038a18d2e in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/deps/./libtorch_cpu.so)
#> frame #11: _lantern_reflection_pad1d_tensor_intarrayref + 0x58 (0x7f904e213198 in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/deps/liblantern.so)
#> frame #12: cpp_torch_namespace_reflection_pad1d_self_Tensor_padding_IntArrayRef(Rcpp::XPtr<XPtrTorchTensor, Rcpp::PreserveStorage, &(void Rcpp::standard_delete_finalizer<XPtrTorchTensor>(XPtrTorchTensor*)), false>, std::vector<long, std::allocator<long> >) + 0x6a (0x7f904eb973aa in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/libs/torchpkg.so)
#> frame #13: _torch_cpp_torch_namespace_reflection_pad1d_self_Tensor_padding_IntArrayRef + 0x9c (0x7f904ea6361c in /home/athos/R/x86_64-pc-linux-gnu-library/4.0/torch/libs/torchpkg.so)
#> frame #14: <unknown function> + 0xf5b3c (0x7f90591c7b3c in /usr/lib/R/lib/libR.so)
#> frame #15: <unknown function> + 0xf6086 (0x7f90591c8086 in /usr/lib/R/lib/libR.so)
#> frame #16: <unknown function> + 0x130181 (0x7f9059202181 in /usr/lib/R/lib/libR.so)
#> frame #17: Rf_eval + 0x88 (0x7f905921cd88 in /usr/lib/R/lib/libR.so)
#> frame #18: <unknown function> + 0x14cc4f (0x7f905921ec4f in /usr/lib/R/lib/libR.so)
#> frame #19: Rf_applyClosure + 0x1a2 (0x7f905921fb42 in /usr/lib/R/lib/libR.so)
#> frame #20: Rf_eval + 0x2af (0x7f905921cfaf in /usr/lib/R/lib/libR.so)
#> frame #21: <unknown function> + 0xc15ad (0x7f90591935ad in /usr/lib/R/lib/libR.so)
#> frame #22: <unknown function> + 0x130181 (0x7f9059202181 in /usr/lib/R/lib/libR.so)
#> frame #23: Rf_eval + 0x88 (0x7f905921cd88 in /usr/lib/R/lib/libR.so)
#> frame #24: <unknown function> + 0x14cc4f (0x7f905921ec4f in /usr/lib/R/lib/libR.so)
#> frame #25: Rf_applyClosure + 0x1a2 (0x7f905921fb42 in /usr/lib/R/lib/libR.so)
#> frame #26: <unknown function> + 0x13a4be (0x7f905920c4be in /usr/lib/R/lib/libR.so)
#> frame #27: Rf_eval + 0x88 (0x7f905921cd88 in /usr/lib/R/lib/libR.so)
#> frame #28: <unknown function> + 0x14cc4f (0x7f905921ec4f in /usr/lib/R/lib/libR.so)
#> frame #29: Rf_applyClosure + 0x1a2 (0x7f905921fb42 in /usr/lib/R/lib/libR.so)
#> frame #30: <unknown function> + 0x13a4be (0x7f905920c4be in /usr/lib/R/lib/libR.so)
#> frame #31: Rf_eval + 0x88 (0x7f905921cd88 in /usr/lib/R/lib/libR.so)
#> frame #32: <unknown function> + 0x14cc4f (0x7f905921ec4f in /usr/lib/R/lib/libR.so)
#> frame #33: Rf_applyClosure + 0x1a2 (0x7f905921fb42 in /usr/lib/R/lib/libR.so)
#> frame #34: <unknown function> + 0x13a4be (0x7f905920c4be in /usr/lib/R/lib/libR.so)
#> frame #35: Rf_eval + 0x88 (0x7f905921cd88 in /usr/lib/R/lib/libR.so)
#> frame #36: <unknown function> + 0x14cc4f (0x7f905921ec4f in /usr/lib/R/lib/libR.so)
#> frame #37: Rf_applyClosure + 0x1a2 (0x7f905921fb42 in /usr/lib/R/lib/libR.so)
#> frame #38: <unknown function> + 0x13a4be (0x7f905920c4be in /usr/lib/R/lib/libR.so)
#> frame #39: Rf_eval + 0x88 (0x7f905921cd88 in /usr/lib/R/lib/libR.so)
#> frame #40: <unknown function> + 0x14cc4f (0x7f905921ec4f in /usr/lib/R/lib/libR.so)
#> frame #41: Rf_applyClosure + 0x1a2 (0x7f905921fb42 in /usr/lib/R/lib/libR.so)
#> frame #42: <unknown function> + 0x13a4be (0x7f905920c4be in /usr/lib/R/lib/libR.so)
#> frame #43: Rf_eval + 0x88 (0x7f905921cd88 in /usr/lib/R/lib/libR.so)
#> frame #44: <unknown function> + 0x14cc4f (0x7f905921ec4f in /usr/lib/R/lib/libR.so)
#> frame #45: Rf_applyClosure + 0x1a2 (0x7f905921fb42 in /usr/lib/R/lib/libR.so)
#> frame #46: <unknown function> + 0x13a4be (0x7f905920c4be in /usr/lib/R/lib/libR.so)
#> frame #47: Rf_eval + 0x88 (0x7f905921cd88 in /usr/lib/R/lib/libR.so)
#> frame #48: <unknown function> + 0x14cc4f (0x7f905921ec4f in /usr/lib/R/lib/libR.so)
#> frame #49: Rf_applyClosure + 0x1a2 (0x7f905921fb42 in /usr/lib/R/lib/libR.so)
#> frame #50: Rf_eval + 0x2af (0x7f905921cfaf in /usr/lib/R/lib/libR.so)
#> frame #51: <unknown function> + 0x14e953 (0x7f9059220953 in /usr/lib/R/lib/libR.so)
#> frame #52: Rf_eval + 0x57b (0x7f905921d27b in /usr/lib/R/lib/libR.so)
#> frame #53: <unknown function> + 0x14cc4f (0x7f905921ec4f in /usr/lib/R/lib/libR.so)
#> frame #54: Rf_applyClosure + 0x1a2 (0x7f905921fb42 in /usr/lib/R/lib/libR.so)
#> frame #55: Rf_eval + 0x2af (0x7f905921cfaf in /usr/lib/R/lib/libR.so)
#> frame #56: <unknown function> + 0x14fef2 (0x7f9059221ef2 in /usr/lib/R/lib/libR.so)
#> frame #57: Rf_eval + 0x57b (0x7f905921d27b in /usr/lib/R/lib/libR.so)
#> frame #58: <unknown function> + 0x14e953 (0x7f9059220953 in /usr/lib/R/lib/libR.so)
#> frame #59: Rf_eval + 0x57b (0x7f905921d27b in /usr/lib/R/lib/libR.so)
#> frame #60: <unknown function> + 0x14cc4f (0x7f905921ec4f in /usr/lib/R/lib/libR.so)
#> frame #61: Rf_applyClosure + 0x1a2 (0x7f905921fb42 in /usr/lib/R/lib/libR.so)
#> frame #62: Rf_eval + 0x2af (0x7f905921cfaf in /usr/lib/R/lib/libR.so)
#> frame #63: <unknown function> + 0x14fef2 (0x7f9059221ef2 in /usr/lib/R/lib/libR.so)

output = wavernn(waveform, specgram)

#> Error in wavernn(waveform, specgram): object 'specgram' not found

# output shape:  (n_batch, n_channel, (n_time - kernel_size + 1) * hop_length, n_classes)